+++ src/shared/components/palette/case-link/components/link-cases/link-cases.component.spec.ts
@@ -24,7 +24,9 @@ describe('LinkCasesComponent', () => {
       reasons: [],
       createdDateTime: '',
       caseType: 'SSCS',
+      caseTypeDescription: 'SSCS case type',
       caseState: 'state',
+      caseStateDescription: 'state description',
       caseService: 'Tribunal',
       caseName: 'SSCS 2.1'
   }];