+++ src/shared/components/palette/case-link/components/unlink-cases/unlink-cases.component.spec.ts
@@ -81,7 +81,9 @@ describe('UnLinkCasesComponent', () => {
       reasons: [],
       createdDateTime: '',
       caseType: 'SSCS',
+      caseTypeDescription: 'SSCS case type',
       caseState: 'state',
+      caseStateDescription: 'state description',
       caseService: 'Tribunal',
       caseName: 'SSCS 2.1'
     },
@@ -90,7 +92,9 @@ describe('UnLinkCasesComponent', () => {
       reasons: [],
       createdDateTime: '',
       caseType: 'SSCS',
+      caseTypeDescription: 'SSCS case type',
       caseState: 'state',
+      caseStateDescription: 'state description',
       caseService: 'Tribunal',
       caseName: 'SSCS 2.1'
     }